refactor: Moved MixerConnections to seperate folder. refactor: Moved Express server to seperate module refactor: Clean up of Global vars
