+++ server/utils/mixerConnections/EmberMixerConnection.ts
@@ -1,14 +1,14 @@
 const DeviceTree = require('emberplus')
-import { store, state } from '../reducers/store'
-import { huiRemoteConnection } from '../mainClasses'
+import { store, state } from '../../reducers/store'
+import { huiRemoteConnection } from '../../mainClasses'
 
 //Utils:
-import { IMixerProtocol } from '../constants/MixerProtocolInterface';
-import { IStore } from '../reducers/indexReducer';
+import { IMixerProtocol } from '../../constants/MixerProtocolInterface';
+import { IStore } from '../../reducers/indexReducer';
 import { 
     SET_FADER_LEVEL, 
     SET_CHANNEL_LABEL 
-} from '../reducers/faderActions'
+} from '../../reducers/faderActions'
 
 
 export class EmberMixerConnection {
