+++ server/utils/mixerConnections/SSLMixerConnection.ts
@@ -1,19 +1,19 @@
 //Node Modules:
 const net = require('net')
-import { store, state } from '../reducers/store'
-import { huiRemoteConnection } from '../mainClasses'
+import { store, state } from '../../reducers/store'
+import { huiRemoteConnection } from '../../mainClasses'
 
 //Utils:
-import { IMixerProtocol } from '../constants/MixerProtocolInterface'
-import { IStore } from '../reducers/indexReducer'
-import { SET_OUTPUT_LEVEL } from '../reducers/channelActions'
+import { IMixerProtocol } from '../../constants/MixerProtocolInterface'
+import { IStore } from '../../reducers/indexReducer'
+import { SET_OUTPUT_LEVEL } from '../../reducers/channelActions'
 import { 
     SET_FADER_LEVEL,
     TOGGLE_PGM,
     SET_MUTE
- } from  '../reducers/faderActions'
-import { SET_MIXER_ONLINE } from '../reducers/settingsActions';
-import { logger } from './logger'
+ } from  '../../reducers/faderActions'
+import { SET_MIXER_ONLINE } from '../../reducers/settingsActions';
+import { logger } from '../logger'
 
 export class SSLMixerConnection {
     mixerProtocol: IMixerProtocol;
